third_party/blink/web_tests/fast/loader/about-blank-reload.html

<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<iframe id="i" src="about:blank"></iframe>
<script>

async_test(t => {
  i.onload = () => {
    assert_equals(i.contentDocument.body.textContent, "");
    t.done();
  }
  i.contentWindow.location.reload();
  i.contentDocument.body.textContent = 'this text should be overwritten by reload';
});
</script>
